Fix the runner
